perf: organize shared actions
