+++ src/shared/actions/Testimony/getTestimonyData.ts
@@ -1,4 +1,4 @@
-"use server"
+"use server";
 
 import prisma from "@/lib/prisma";
